[FEAT/#97]: 소상공인, 도매업체 공동구매 현황 컴포넌트 분리
